rename v1 test module to integration
