+++ crates/wasmi/tests/integration/call_hook.rs
