+++ crates/wasmi/tests/integration/fuel_consumption.rs
