+++ crates/wasmi/tests/integration/fuel_metering.rs
