+++ crates/wasmi/tests/integration/func.rs
