+++ crates/wasmi/tests/integration/host_call_compilation.rs
