+++ crates/wasmi/tests/integration/host_call_instantiation.rs
