+++ crates/wasmi/tests/integration/host_calls_wasm.rs
