+++ crates/wasmi/tests/integration/resource_limiter.rs
