+++ crates/wasmi/tests/integration/resumable_call.rs
